# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.createQuestionRequest:
    properties:
      analysis:
        maxLength: 2000
        type: string
      answer:
        items:
          type: integer
        type: array
      content:
        maxLength: 2000
        minLength: 2
        type: string
      default_score:
        maximum: 100
        minimum: 1
        type: integer
      difficulty:
        maximum: 5
        minimum: 1
        type: integer
      options:
        items:
          $ref: '#/definitions/model.QuestionOption'
        maxItems: 10
        type: array
      tags:
        maxLength: 255
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.QuestionType'
        enum:
        - single
        - multiple
        - judge
    required:
    - answer
    - content
    - type
    type: object
...
  handler.updateQuestionRequest:
    properties:
      analysis:
        maxLength: 2000
        type: string
      answer:
        items:
          type: integer
        type: array
      content:
        maxLength: 2000
        minLength: 2
        type: string
      default_score:
        maximum: 100
        minimum: 1
        type: integer
      difficulty:
        maximum: 5
        minimum: 1
        type: integer
      options:
        items:
          $ref: '#/definitions/model.QuestionOption'
        maxItems: 10
        type: array
      status:
        allOf:
        - $ref: '#/definitions/model.QuestionStatus'
        enum:
        - enabled
        - disabled
      tags:
        maxLength: 255
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.QuestionType'
        enum:
        - single
        - multiple
        - judge
    type: object
...
  hzycoder_com_lion_internal_model_dto_response.QueryUser:
    properties:
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/model.Role'
      username:
        type: string
    type: object
  hzycoder_com_lion_internal_model_dto_response.QuestionItem:
    properties:
      analysis:
        type: string
      answer:
        items:
          type: integer
        type: array
      content:
        type: string
      created_at:
        type: string
      default_score:
        type: integer
      difficulty:
        type: integer
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/model.QuestionOption'
        type: array
      status:
        $ref: '#/definitions/model.QuestionStatus'
      tags:
        type: string
      type:
        $ref: '#/definitions/model.QuestionType'
    type: object
  hzycoder_com_lion_internal_model_dto_response.QuestionPage:
    properties:
      list:
        items:
          $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QuestionItem'
        type: array
      total:
        type: integer
    type: object
...
  model.QuestionOption:
    properties:
      key:
        type: string
      text:
        type: string
    type: object
  model.QuestionStatus:
    enum:
    - enabled
    - disabled
    type: string
    x-enum-varnames:
    - QuestionStatusEnabled
    - QuestionStatusDisabled
  model.QuestionType:
    enum:
    - single
    - multiple
    - judge
    type: string
    x-enum-varnames:
    - QuestionTypeSingle
    - QuestionTypeMultiple
    - QuestionTypeJudge
  model.Role:
    enum:
    - guest
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleGuest
    - RoleAdmin
    - RoleMember
  response.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 试卷系统后端服务：题库管理、手动组卷与发布、在线答题、自动评分、成绩查询与统计。
  title: Lion 试卷系统 API
  version: "1.0"
paths:
...
  /v1/admin/questions:
    get:
      description: 支持题型/状态/难度/标签筛选与关键词搜索（匹配题干）
      parameters:
      - description: 题型：single multiple judge
        enum:
        - single
        - multiple
        - judge
        in: query
        name: type
        type: string
      - description: 状态：enabled disabled
        enum:
        - enabled
        - disabled
        in: query
        name: status
        type: string
      - description: 难度 1-5
        in: query
        name: difficulty
        type: integer
      - description: 标签（逗号分隔标签之一即命中）
        in: query
        name: tag
        type: string
      - description: 题干关键词
        in: query
        name: keyword
        type: string
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认 20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QuestionPage'
              type: object
      security:
      - BearerAuth: []
      summary: 题目分页列表
      tags:
      - 题库管理
    post:
      consumes:
      - application/json
      description: 创建单选/多选/判断题；tags 为逗号分隔字符串；answer 为 JSON（单选/判断传字符串，多选传字符串数组）
      parameters:
      - description: 题目内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QuestionItem'
              type: object
      security:
      - BearerAuth: []
      summary: 新建题目
      tags:
      - 题库管理
  /v1/admin/questions/{id}:
    delete:
      description: 已被试卷引用的题目不可删除
      parameters:
      - description: 题目 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 删除题目
      tags:
      - 题库管理
    get:
      parameters:
      - description: 题目 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QuestionItem'
              type: object
      security:
      - BearerAuth: []
      summary: 题目详情
      tags:
      - 题库管理
    put:
      consumes:
      - application/json
      description: 全字段可选更新；status 可切换 enabled/disabled
      parameters:
      - description: 题目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 待更新字段
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QuestionItem'
              type: object
      security:
      - BearerAuth: []
      summary: 编辑题目
      tags:
      - 题库管理
...
  /v1/users/{username}:
    get:
      parameters:
      - description: 用户名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QueryUser'
              type: object
      security:
      - BearerAuth: []
      summary: 按用户名查询用户（仅管理员）
      tags:
      - 用户
  /v1/users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/hzycoder_com_lion_internal_model_dto_response.QueryUser'
              type: object
      security:
      - BearerAuth: []
      summary: 当前登录用户信息
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    description: 登录或注册接口签发的 JWT，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"